+++ src/main/java/com/samyak/swastik/domain/LotCategory.java
@@ -47,13 +47,13 @@ public class LotCategory implements Serializable {
 	private String modifeidMachineName;
 
 	@Column(name = "SR_NO")
-	private Integer srNo;
+	private Double srNo;
 
 	@Column(name = "ACTIVE")
 	private Boolean active;
 
 	@Column(name = "YEAR_END_ID")
-	private Integer yearEndId;
+	private Double yearEndId;
 
 	@Column(name = "LOT_CATEGORY_NAME_NATIVE")
 	private String lotCategoryNameNative;
@@ -130,11 +130,11 @@ public class LotCategory implements Serializable {
 		this.modifeidMachineName = modifeidMachineName;
 	}
 
-	public Integer getSrNo() {
+	public Double getSrNo() {
 		return srNo;
 	}
 
-	public void setSrNo(Integer srNo) {
+	public void setSrNo(Double srNo) {
 		this.srNo = srNo;
 	}
 
@@ -146,11 +146,11 @@ public class LotCategory implements Serializable {
 		this.active = active;
 	}
 
-	public Integer getYearEndId() {
+	public Double getYearEndId() {
 		return yearEndId;
 	}
 
-	public void setYearEndId(Integer yearEndId) {
+	public void setYearEndId(Double yearEndId) {
 		this.yearEndId = yearEndId;
 	}
